refactor: setup property
